added "?? []" behind every session since no guarantee there will be watchlist data from session hook however it didn't build the other day but tdy it build?? magic didn't even do anything
