+++ app/watch-list/action.js
@@ -2,7 +2,7 @@
 import { ANIME } from '@consumet/extensions';
 import clientPromise from "@/lib/mongodb";
 import { getServerSession } from "next-auth/next";
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+import { authOptions } from '@/lib/authOptions';
 
 export const fetchWatchList = async () => {
   try {
