+++ app/watch-list/page.tsx
@@ -2,7 +2,7 @@ import { fetchWatchList } from "./action";
 import Image from "next/image";
 import Link from "next/link";
 import { getServerSession } from "next-auth/next";
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+import { authOptions } from "@/lib/authOptions";
 import WatchListButton from "@/components/WatchListButton/WatchListButton";
 
 export default async function WatchList() {
