+++ components/WatchListButton/action.js
@@ -2,7 +2,7 @@
 
 import clientPromise from '@/lib/mongodb';
 import { getServerSession } from 'next-auth';
-import { authOptions } from '@/app/api/auth/[...nextauth]/route';
+import { authOptions } from '@/lib/authOptions';
 
 export const handleAddToList = async (id, title) => {
   try {
