+++ package.json
@@ -16,10 +16,10 @@
     "bcryptjs": "^2.4.3",
     "jsonwebtoken": "^9.0.2",
     "mongodb": "^6.8.0",
-    "next": "14.0.4",
+    "next": "^14.2.7",
     "next-auth": "^4.24.7",
-    "react": "^18",
-    "react-dom": "^18",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-player": "^2.14.1"
   },
   "devDependencies": {
@@ -29,7 +29,7 @@
     "@types/react-dom": "^18",
     "autoprefixer": "^10.0.1",
     "eslint": "^8",
-    "eslint-config-next": "14.0.4",
+    "eslint-config-next": "^14.2.7",
     "postcss": "^8",
     "tailwindcss": "^3.3.0",
     "typescript": "^5"
